# src/main/resources/swagger/swagger-workstation.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: API Workstation
  description: 'Hackathon'
  contact:
    name: Team 3 - GFT Brasil
    email: andu@gft.com
host: gft.com.br
basePath: /gft/workstation/v1/office/
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /workstation:
    post:
      tags:
        - Workstation
      summary: Create new workstation
      description: 'Create new workstation.'
      operationId: createWorkstation
      produces:
        - application/json
      parameters:
        - in: body
          name: newWorkstation
          description: data of the new workstation
          required: true
          schema:
            $ref: '#/definitions/Workstation'

      responses:
        '201':
          description: Create new resource
          schema:
            $ref: '#/definitions/WorkstationCode'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
      security:
        - OauthSecurity:
            - "bearer"
    get:
      tags:
        - Workstation
      summary: Gets all workstation
      description: 'get a workstation'
      operationId: getWorkstation
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/WorkstationsResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /workstation/{id}:
    get:
      tags:
        - Workstation
      summary: Gets a workstation
      description: 'get a workstation'
      operationId: getWorkstationById
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of workstation
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Workstation'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /workstation/{id}/blocked:
    patch:
      tags:
        - Workstation
      summary: Patch a workstation
      description: 'get a workstation'
      operationId: blockedWorkstation
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of workstation
          required: true
          type: string
          x-example: '000001'
      responses:
        '204':
          description: Ok
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /workstation/{id}/unblocked:
    patch:
      tags:
        - Workstation
      summary: Patch a workstation
      description: 'get a workstation'
      operationId: unblockedWorkstation
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of workstation
          required: true
          type: string
          x-example: '000001'
      responses:
        '204':
          description: Ok
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /user:
    post:
      tags:
        - User
      summary: Create new user
      description: 'Create new user.'
      operationId: createUser
      produces:
        - application/json
      parameters:
        - in: body
          name: newUser
          description: data of the new User
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:

        '201':
          description: Create new resource
          schema:
            $ref: '#/definitions/UserCode'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
      security:
        - OauthSecurity:
            - "bearer"
    get:
      tags:
        - User
      summary: Gets all users
      description: 'get a user'
      operationId: getUsers
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/UsersResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /user/{id}:
    get:
      tags:
        - User
      summary: Gets a user
      description: 'get a user'
      operationId: getUser
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of user
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/UserResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'

    put:
      tags:
        - User
      summary: Update a user
      description: 'put a user'
      operationId: updateUser
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of the user
          required: true
          type: string
          x-example: 'XPTO'
        - name: User
          in: body
          description: data of the User
          required: true
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'

    delete:
      tags:
        - User
      summary: Gets a user
      description: 'get a user'
      operationId: deleteUser
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of user
          required: true
          type: string
          x-example: '000001'
      responses:
        '204':
          description: No Content
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
securityDefinitions:
  OauthSecurity:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    tokenUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    scopes:
      bearer: Client authorization token - Bearer JWT
security:
  - OauthSecurity:
      - bearer
definitions:
...
  User:
    description: Object User
    type: object
    required:
      - name
    properties:
      name:
        type: string
        example: string
        description: Name of User
      user:
        type: string
        example: andu
        description: Unique identifier of the User
      email:
        type: string
        example: string
        description: Email
      status:
        type: string
        enum: [ACTIVE, INACTIVE]
      type:
        type: string
        example: ADMIN
        description: ADMIN | CLIENT
        enum: [ADMIN, CLIENT]
  UsersResponse:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/UserResponse'
  UserResponse:
    description: Object User
    type: object
    required:
      - name
    properties:
      id:
        type: string
        example: 001
      name:
        type: string
        example: string
        description: Name of product
      user:
        type: string
        example: andu
        description: Unique identifier of the User
      email:
        type: string
        example: string
        description: Name of product
      createdDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
      updateDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
      status:
        type: string
        enum: [ativo, inativo]
      type:
        type: string
        example: Tipo de usuario
        description: Admin | Cliente
        enum: [admin, cliente]
  UpdateUser:
    description: Object Product
    type: object
    required:
      - name
      - description
    properties:
      name:
        type: string
        example: string
        description: Name of product
      user:
        type: string
        example: andu
        description: Unique identifier of the User
      email:
        type: string
        example: string
        description: Name of product
      createdDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Start Date of the validity of the product.
      status:
        type: string
        enum: [ativo, inativo]
      type:
        type: string
        example: Tipo de usuario
        description: Admin | Cliente
        enum: [admin, cliente]
  Workstation:
    description: Workstation
    type: object
    required:
      - name
    properties:
      id:
        type: string
        example: 001
        description: Name of Workstation
      idOffice:
        type: string
      name:
        type: string
        example: string
        description: Name of Workstation
      description:
        type: string
        example: Fica perto do banheiro
        description: Name of Workstation
      createdDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Start Date of the validity of the workstation.
      group:
        type: string
        example: Bloco-X22
      status:
        type: string
        enum: [ACTIVE, INACTIVE]
      blocked:
        type: string
        enum: [YES, NO]
      typeWorkstation:
        type: string
        enum: [DESKTOP, DOCK_STATION, PERIFERICOS]
  WorkstationsResponse:
    type: object
    properties:
      workstations:
        type: array
        items:
          $ref: '#/definitions/Workstation'
  WorkstationCode:
    description: Workstation data
    type: object
    required:
      - id
    properties:
      id:
        type: string
        example: 001
        description: Unique identifier of the Workstation
...
  UserCode:
    description: User id
    type: object
    required:
      - id
    properties:
      id:
        type: string
        example: 001
        description: Unique identifier of the User
  ErrorModel:
    type: object
    required:
      - codeBuilderComplete
      - message
    properties:
      code:
        type: string
      message:
        type: string